refactor: molecules naming convention
